+++ ts/packages/core/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @composio/core
 
+## 0.1.36-next.13
+
+### Patch Changes
+
+- 77e96e4: Fix JSON Schema to Zod Parsing
+- Updated dependencies [77e96e4]
+  - @composio/json-schema-to-zod@0.1.10-next.5
+
+## 0.1.36-next.12
+
+### Patch Changes
+
+- Update deps
+
+## 0.1.36-next.11
+
+### Patch Changes
+
+- Fix proxy execute params and bump langchain packages
+
 ## 0.1.36-next.10
 
 ### Patch Changes