+++ ts/packages/providers/cloudflare/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @composio/cloudflare
 
+## 0.1.36-next.13
+
+### Patch Changes
+
+- 77e96e4: Fix JSON Schema to Zod Parsing
+- Updated dependencies [77e96e4]
+  - @composio/core@0.1.36-next.13
+
+## 0.1.36-next.12
+
+### Patch Changes
+
+- Update deps
+- Updated dependencies
+  - @composio/core@0.1.36-next.12
+
+## 0.1.36-next.11
+
+### Patch Changes
+
+- Fix proxy execute params and bump langchain packages
+- Updated dependencies
+  - @composio/core@0.1.36-next.11
+
 ## 0.1.36-next.10
 
 ### Patch Changes